AI-382: move AgentsWorkflowError to its own module
The error type lived in workflow.py, which imports TemporalSandboxClient
from the sandbox client module. The sandbox client's host-path-grant
rejection needed the error, so it reached for it with a deferred
function-body import to break the cycle. Moving the class to a private
_errors module lets every importer take it at module level.

workflow.py re-exports the name so
temporalio.contrib.openai_agents.workflow.AgentsWorkflowError keeps
resolving for existing callers.

The docstring now covers both origins: an agents SDK error rewrapped, and
the plugin rejecting an unsupported configuration.
